Updated a ton of files
MAJOR WIP:
GameUI has troubles, but I updated Unit, Player, Bomberman, and tons of Engine classes. 

The Unit selecting with the DxMouse is a bit less messy now, since it uses the Rect contains method instead of checking hard coded boundaries

The last thing that needs to be fixed on Units are the collisions. Currently, Units can be selected and moved, but they get stuck in collidable objects and then can't get out. I'm guessing that the collision checking isn't being done as soon as the Unit moves